Package all ui/*.ts classes in one .ts file.
