+++ src/flash/references.ts
@@ -382,24 +382,25 @@
 
 ///<reference path='trace/Trace.ts' />
 
-///<reference path='ui/ContextMenu.ts' />
-///<reference path='ui/ContextMenuBuiltInItems.ts' />
-///<reference path='ui/ContextMenuClipboardItems.ts' />
-///<reference path='ui/ContextMenuItem.ts' />
-///<reference path='ui/GameInput.ts' />
-///<reference path='ui/GameInputControl.ts' />
-///<reference path='ui/GameInputControlType.ts' />
-///<reference path='ui/GameInputDevice.ts' />
-///<reference path='ui/GameInputFinger.ts' />
-///<reference path='ui/GameInputHand.ts' />
-// ///<reference path='ui/KeyLocation.ts' />
-///<reference path='ui/Keyboard.ts' />
-// ///<reference path='ui/KeyboardType.ts' />
-///<reference path='ui/Mouse.ts' />
-///<reference path='ui/MouseCursor.ts' />
-///<reference path='ui/MouseCursorData.ts' />
-///<reference path='ui/Multitouch.ts' />
-///<reference path='ui/MultitouchInputMode.ts' />
+///<reference path='ui/ui.ts' />
+// ///<reference path='ui/ContextMenu.ts' />
+// ///<reference path='ui/ContextMenuBuiltInItems.ts' />
+// ///<reference path='ui/ContextMenuClipboardItems.ts' />
+// ///<reference path='ui/ContextMenuItem.ts' />
+// ///<reference path='ui/GameInput.ts' />
+// ///<reference path='ui/GameInputControl.ts' />
+// ///<reference path='ui/GameInputControlType.ts' />
+// ///<reference path='ui/GameInputDevice.ts' />
+// ///<reference path='ui/GameInputFinger.ts' />
+// ///<reference path='ui/GameInputHand.ts' />
+// // ///<reference path='ui/KeyLocation.ts' />
+// ///<reference path='ui/Keyboard.ts' />
+// // ///<reference path='ui/KeyboardType.ts' />
+// ///<reference path='ui/Mouse.ts' />
+// ///<reference path='ui/MouseCursor.ts' />
+// ///<reference path='ui/MouseCursorData.ts' />
+// ///<reference path='ui/Multitouch.ts' />
+// ///<reference path='ui/MultitouchInputMode.ts' />
 
 ///<reference path='utils/Endian.ts' />
 ///<reference path='utils/IDataInput2.ts' />
